+++ admin/controllers/FMControllerSelect_data_from_db.php
@@ -20,8 +20,8 @@ class FMControllerSelect_data_from_db {
   ////////////////////////////////////////////////////////////////////////////////////////
   public function execute() {
     $task = ((isset($_GET['task'])) ? esc_html($_GET['task']) : 0);
-    $id = ((isset($_GET['id'])) ? esc_html($_GET['id']) : 0);
-    $form_id = ((isset($_GET['form_id'])) ? esc_html($_GET['form_id']) : 0);
+    $id = ((isset($_GET['id'])) ? (int)esc_html($_GET['id']) : 0);
+    $form_id = ((isset($_GET['form_id'])) ? (int)esc_html($_GET['form_id']) : 0);
     $field_id = ((isset($_GET['field_id'])) ? esc_html($_GET['field_id']) : 0);
 	//var_dump($form_id);
     $field_type = ((isset($_GET['field_type'])) ? esc_html($_GET['field_type']) : 0);
@@ -48,7 +48,7 @@ class FMControllerSelect_data_from_db {
 
     require_once WD_FM_DIR . "/admin/views/FMViewSelect_data_from_db.php";
     $view = new FMViewSelect_data_from_db($model);
-    $view->db_tables($form_id,$field_type);
+    $view->db_tables((int)$form_id,$field_type);
   }
    public function db_table_struct_select($form_id,$field_type) {
     require_once WD_FM_DIR . "/admin/models/FMModelSelect_data_from_db.php";
@@ -56,7 +56,7 @@ class FMControllerSelect_data_from_db {
 
     require_once WD_FM_DIR . "/admin/views/FMViewSelect_data_from_db.php";
     $view = new FMViewSelect_data_from_db($model);
-    $view->db_table_struct_select($form_id,$field_type);
+    $view->db_table_struct_select((int)$form_id,$field_type);
   }
 
   ////////////////////////////////////////////////////////////////////////////////////////
